requirements.txt
streamlit
langchain
langchain-community
openai
langchain-openai
sentence_transformers
qdrant-client
python-dotenv
langchainhub